// App/src/config/constants/types.ts
export enum PoolIds {
  poolBasic = 'poolBasic',
  poolUnlimited = 'poolUnlimited',
}
// ...
interface MintingPoolInfo {
  saleAmount: string
  distributionRatio: number // Range [0-1]
}

//CC
export interface Minting {
  id: string
  stake_pid?: number
  name: string
  isActive: boolean
  status: string
  address: string
  symbol: string
  totalSupply: number
  avatar: string
  banner: {large: string, small: string}
  sampleNftImage?: {tokenId: number, image: string, link: string}
  showCase?: any[]
  faq?: {title: string, description: string[]}[]
  currency: Token
  token: Token
  releaseBlockNumber: number
  lastPrice?: number
  description?: string
  articleUrl?: string
  twitterUrl?: string
  telegramUrl?: string
  openSeaUrl?: string
  cmcUrl?: string
  discordUrl?: string
  version: number
  [PoolIds.poolBasic]?: MintingPoolInfo
  [PoolIds.poolUnlimited]: MintingPoolInfo
}